+++ CHANGELOG.md
@@ -16,6 +16,7 @@ A preview of the next release can be installed from
 - `dissoc` on record with non-record field should return map instead of record
 - Bump edamame to `1.5.34`
 - Bump `core.rrb-vector` to `0.2.0`
+- Migrate detecting of executable name for self-executing uberjar executable from `ProcessHandle` to to native image `ProcessInfo` to avoid sandbox errors
 
 ## 1.12.209 (2025-10-02)